moved db connection function to config folder
