+++ src/index.js
@@ -1,7 +1,6 @@
-require("dotenv").config();
+const connectToDatabase = require("./config/connection");
 
 const express = require("express");
-const mongoose = require("mongoose");
 
 const routes = require("./routes");
 
@@ -16,17 +15,7 @@ app.use(routes);
 
 const init = async () => {
   try {
-    const DB_NAME = process.env.DB_NAME;
-    //getting an error with `mongodb://localhost:27017/${DB_NAME} and server doesn't start, so replaced it with `mongodb://0.0.0.0:27017/${DB_NAME}`
-    const MONGODB_URI =
-      process.env.MONGODB_URI || `mongodb://0.0.0.0:27017/${DB_NAME}`;
-
-    const options = {
-      useNewUrlParser: true,
-      useUnifiedTopology: true,
-    };
-
-    await mongoose.connect(MONGODB_URI, options);
+    await connectToDatabase();
 
     app.listen(PORT, () => {
       console.log(`Server running on http://localhost:${PORT}`);
